requirements.txt
numpy==1.19.1
pandas==1.0.5
tqdm==4.48.0
Pillow==8.2.0
scikit_learn==0.24.2
torch==1.8.1
torchvision==0.9.1
